+++ src/store/mod.rs
@@ -1,9 +1,8 @@
+mod memory;
+pub use memory::*;
+
 #[cfg(feature = "redis-store")]
 pub mod redis;
 
-mod memory;
 mod store_trait;
-
-pub use memory::*;
-
 pub use store_trait::*;